requirements.txt
nxtools
htmlmin
jsmin
jinja2
libsass
typer
watchdog
